fix: entities/players now visible through warpzone portals (2 stacked root causes)
The user report "attacks pass through but nothing is visible on the other
side - no players or entities" had two independent causes, isolated with the
scripted dump loop (lookat/force/probe cvars, main-vs-portal comparisons):

1. Godot's FRUSTUM projection (Camera3D.SetFrustum) does not render
   small-AABB instances at all in this scene (its light culler also rejects
   the matrices - prepare_camera spam): items/players never drew while world
   cells (huge AABBs) did. Replaced the window-anchored frustum with a
   PERSPECTIVE exit camera + PROJECTIVE UV sampling: the camera still sits at
   the warp-transformed eye looking along the exit normal, sized to cover the
   window cone; because the view axis is the plane normal, near=planeDist+0.5
   clips the plane-coincident wall face IDENTICALLY to the frustum (the
   window is now an off-center subrect of the image - the quad maps each
   fragment through the exit projection via wz_off/wz_dist/wz_tan uniforms).
   Also kills the residual prepare_camera light-culler spam for good.

2. The player pose-cull (cl_pose_cull, default 1) skips the skeleton bone
   push for models whose VisibleOnScreenNotifier3D says off-screen - which
   only reflects the MAIN camera. A player visible ONLY through a portal
   kept an unpushed skeleton (rest pose, degenerate AABB - the PlayerModel
   comment warns exactly this) and was culled in the portal view. While any
   portal renders, the cull is bypassed (portals are occasional; cost
   reverts to baseline only while one is on screen).

Verified with the dump loop: a roaming bot + dropped weapon now render
through the portal (probe log shows the bot in-frustum, and the texture
shows it). Debug additions: wz_portal_lookat N / N+10 (park the main camera
at a portal's entry / inside its exit room), [portal-ent] in-region entity
probe, per-viewport rendered-object counts in dumps.

2916 tests green.

+++ game/client/EffectSystem.cs
@@ -454,7 +454,7 @@ public partial class EffectSystem : Node3D
     /// <paramref name="color"/> tints the particles (QC eent_net_color override).
     /// Returns the created node (already added to the tree), or null if the name is empty/Null.
     /// </summary>
-    public Node3D? Spawn(string effectName, NVec3 origin, NVec3 velocity = default, int count = 1, Color? color = null)
+    public Node3D? Spawn(string effectName, NVec3 origin, NVec3 velocity = default, float count = 1f, Color? color = null)
     {
         if (string.IsNullOrEmpty(effectName))
             return null;
@@ -1171,7 +1171,7 @@ public partial class EffectSystem : Node3D
     /// effect, or the trail END point for a trail effect (then the segment is origin..velocity).
     /// </summary>
     private Node3D? BuildFromInfo(IReadOnlyList<EffectInfoEmitter> blocks, NVec3 origin, NVec3 velocity,
-        int count, Color? colorOverride, bool isTrail, bool attach = true)
+        float count, Color? colorOverride, bool isTrail, bool attach = true)
     {
         var parent = new Node3D { Name = "fx_info", Position = Coords.ToGodot(origin) };
         bool any = false;
@@ -1343,7 +1343,7 @@ public partial class EffectSystem : Node3D
 
     /// <summary>Build one GPU-particle emitter from a single parsed emitter block (the per-block spawn loop).</summary>
     private GpuParticles3D? BuildInfoBurst(EffectInfoEmitter info, NVec3 originMin, NVec3 originMax,
-        NVec3 emitVel, float traillen, int requestedCount, Color? colorOverride, bool isTrail,
+        NVec3 emitVel, float traillen, float requestedCount, Color? colorOverride, bool isTrail,
         NVec3 fwd, NVec3 right, NVec3 up)
     {
         // DP cl_particles.c:1710-1721: cnt = countabsolute + (pcount*countmultiplier)*quality, plus the
